requirements.txt
aiohttp
python-dotenv
pandas
numpy
aiodns
sqlalchemy
asyncpg
greenlet